refractor: Move election stuff in jobs into their own package
