+++ src/main/kotlin/com/github/recraftedcivilizations/darkcitizens/commands/ShowJobs.kt
@@ -5,13 +5,11 @@ import com.github.darkvanityoflight.recraftedcore.gui.elements.CloseButtonFactor
 import com.github.darkvanityoflight.recraftedcore.utils.itemutils.addLore
 import com.github.darkvanityoflight.recraftedcore.utils.itemutils.getName
 import com.github.darkvanityoflight.recraftedcore.utils.itemutils.setName
-import com.github.recraftedcivilizations.darkcitizens.BukkitWrapper
 import com.github.recraftedcivilizations.darkcitizens.dPlayer.DPlayerManager
 import com.github.recraftedcivilizations.darkcitizens.election.ElectionManager
 import com.github.recraftedcivilizations.darkcitizens.gui.JobItem
-import com.github.recraftedcivilizations.darkcitizens.jobs.ElectedJob
+import com.github.recraftedcivilizations.darkcitizens.jobs.elected.ElectedJob
 import com.github.recraftedcivilizations.darkcitizens.jobs.JobManager
-import org.bukkit.Material
 import org.bukkit.command.Command
 import org.bukkit.command.CommandExecutor
 import org.bukkit.command.CommandSender
