+++ src/main/kotlin/com/github/recraftedcivilizations/darkcitizens/listeners/ElectionTrigger.kt
@@ -3,7 +3,7 @@ package com.github.recraftedcivilizations.darkcitizens.listeners
 import com.github.recraftedcivilizations.darkcitizens.dPlayer.DPlayerManager
 import com.github.recraftedcivilizations.darkcitizens.election.ElectionManager
 import com.github.recraftedcivilizations.darkcitizens.events.JobLeaveEvent
-import com.github.recraftedcivilizations.darkcitizens.jobs.ElectedJob
+import com.github.recraftedcivilizations.darkcitizens.jobs.elected.ElectedJob
 import com.github.recraftedcivilizations.darkcitizens.jobs.JobManager
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
