+++ src/test/kotlin/com/github/recraftedcivilizations/jobs/JobFactoryTest.kt
@@ -2,7 +2,7 @@ package com.github.recraftedcivilizations.jobs
 
 
 import com.github.recraftedcivilizations.darkcitizens.dPlayer.DPlayerManager
-import com.github.recraftedcivilizations.darkcitizens.jobs.ElectedJob
+import com.github.recraftedcivilizations.darkcitizens.jobs.elected.ElectedJob
 import com.github.recraftedcivilizations.darkcitizens.jobs.Job
 import com.github.recraftedcivilizations.darkcitizens.jobs.JobFactory
 import com.github.recraftedcivilizations.darkcitizens.jobs.JobManager
